Update Client form
