+++ src/CSBill/ClientBundle/Controller/AjaxController.php
@@ -14,8 +14,8 @@ namespace CSBill\ClientBundle\Controller;
 use CSBill\ClientBundle\Entity\Address;
 use CSBill\ClientBundle\Entity\Client;
 use CSBill\ClientBundle\Entity\Contact;
-use CSBill\ClientBundle\Form\Contact as ContactType;
 use CSBill\ClientBundle\Form\Type\AddressType;
+use CSBill\ClientBundle\Form\Type\ContactType;
 use CSBill\CoreBundle\Controller\BaseController;
 use JMS\Serializer\SerializationContext;
 use Symfony\Component\HttpFoundation\Request;
