+++ src/CSBill/ClientBundle/Controller/DefaultController.php
@@ -12,7 +12,7 @@
 namespace CSBill\ClientBundle\Controller;
 
 use CSBill\ClientBundle\Entity\Client;
-use CSBill\ClientBundle\Form\Client as ClientForm;
+use CSBill\ClientBundle\Form\Type\ClientType;
 use CSBill\ClientBundle\Model\Status;
 use CSBill\CoreBundle\Controller\BaseController;
 use CSBill\InvoiceBundle\Model\Graph;
@@ -42,7 +42,7 @@ class DefaultController extends BaseController
     {
         $client = new Client();
 
-        $form = $this->createForm(new ClientForm(), $client);
+	$form = $this->createForm(ClientType::class, $client);
         $form->handleRequest($request);
 
         if ($form->isValid()) {
@@ -73,7 +73,7 @@ class DefaultController extends BaseController
         $originalAddresses = $client->getAddresses()->toArray();
         $originalContactsDetails = $this->getClientContactDetails($request, $client);
 
-        $form = $this->createForm(ClientForm::class, $client);
+	$form = $this->createForm(ClientType::class, $client);
 
         $form->handleRequest($request);
