+++ src/CSBill/ClientBundle/Entity/Client.php
@@ -471,7 +471,7 @@ class Client
      */
     public function __toString()
     {
-        return $this->name;
+	return $this->name ?? '';
     }
 
     /**
