--- src/CSBill/ClientBundle/Form/Client.php
@@ -1,84 +1,0 @@
-<?php
-
-/*
- * This file is part of CSBill project.
- *
- * (c) 2013-2016 Pierre du Plessis <info@customscripts.co.za>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
-namespace CSBill\ClientBundle\Form;
-
-use CSBill\ClientBundle\Form\Type\AddressType;
-use CSBill\ClientBundle\Form\Type\ContactType;
-use CSBill\CoreBundle\Form\Type\Select2;
-use Money\Currency;
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-class Client extends AbstractType
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder->add('name');
-        $builder->add('website');
-        $builder->add(
-            'currency',
-            Select2::class,
-            [
-                'attr' => ['class' => 'select2'],
-                'empty_value' => 'client.form.currency.empty_value',
-                'choices' => array_flip(Currency::getCurrencies()),
-                'choices_as_values' => true,
-            ]
-        );
-
-        $builder->add(
-            'contacts',
-            ContactType::class,
-            [
-                'entry_type' => Contact::class,
-                'allow_add' => true,
-                'allow_delete' => true,
-                'by_reference' => false,
-                'prototype' => true,
-                'prototype_name' => '__contact_prototype__',
-            ]
-        );
-
-        $builder->add(
-            'addresses',
-            CollectionType::class,
-            [
-                'entry_type' => AddressType::class,
-                'allow_add' => true,
-                'allow_delete' => true,
-                'by_reference' => false,
-                'required' => false,
-            ]
-        );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefaults(['data_class' => 'CSBill\ClientBundle\Entity\Client']);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getBlockPrefix()
-    {
-        return 'client';
-    }
-}
