+++ src/CSBill/ClientBundle/Resources/config/services.yml
@@ -2,7 +2,7 @@ parameters:
     csbill_client.menu.class: CSBill\ClientBundle\Menu\Builder
     csbill_client.contact_type.repository.class: CSBill\ClientBundle\Repository\ContactTypeRepository
     csbill_client.contact_types.twig.extension.class: CSBill\ClientBundle\Twig\Extension\ContactTypesExtension
-    csbill_client.contact.form.type.class: CSBill\ClientBundle\Form\Contact
+
 services:
     csbill_client.menu:
         class: '%csbill_client.menu.class%'
@@ -14,18 +14,16 @@ services:
             - { name: cs_core.menu, menu: clients_edit, method: clientsMenu }
 
     csbill_client.contact.form.type:
-        class: '%csbill_client.contact.form.type.class%'
-        arguments:
-            - '@csbill_client.contact_type.values'
+	class: CSBill\ClientBundle\Form\Type\ContactType
         tags:
             - { name: form.type }
 
-    csbill_client.contact_types.twig.extension:
-        class: '%csbill_client.contact_types.twig.extension.class%'
+    csbill_client.contact_detail.form.type:
+	class: CSBill\ClientBundle\Form\Type\ContactDetailType
         arguments:
-            - '@doctrine'
+	    - '@csbill_client.contact_type.values'
         tags:
-            - { name: twig.extension }
+	    - { name: form.type }
 
     csbill_client.contact_type.repository:
         class: '%csbill_client.contact_type.repository.class%'
