+++ src/CSBill/ClientBundle/Resources/views/Form/contact_details.html.twig
@@ -11,7 +11,7 @@
     <div class="input-group-btn input-group-select">
         <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown">
             <span class="concept">
-                {% for type in contactTypes() if type.id == form.type.vars.value %}
+		{% for type in contactTypes if type.id == form.type.vars.value %}
                     {{ type.name|title }}
                 {% else %}
                     {{ 'client.contact.details.choose'|trans }}
@@ -19,7 +19,7 @@
             </span> <span class="caret"></span>
         </button>
         <ul class="dropdown-menu" role="menu">
-            {% for type in contactTypes() %}
+	    {% for type in contactTypes %}
                 <li>
                     <a href="#" data-value="{{ type.id }}" data-type="{{ type.name }}">
                         {{ type.name|title }}
@@ -39,5 +39,7 @@
     </span>
 </div>
 
-{{ form_errors(form.value) }}
-{{ form_errors(form) }}
+<div class="form-group has-error">
+    {{ form_errors(form.value) }}
+    {{ form_errors(form) }}
+</div>
